1. Allow all policy actions to delay commiting. 2. Make policy actions all commit by default. 3. Provide a syncAllPolicies action to sync policies if they're dirty. 4. Split select/deselect out of tools and into their own action modules. 5. Clean up tools.select to transfer to tool select/deselect handlers.
