+++ src/js/actions/index.js
@@ -48,6 +48,14 @@ define(function (require, exports, module) {
         shapes: require("./shapes"),
         shortcuts: require("./shortcuts"),
         superselect: require("./superselect"),
+        toolEllipse: require("./tool/ellipse"),
+        toolPen: require("./tool/pen"),
+        toolRectangle: require("./tool/rectangle"),
+        toolSampler: require("./tool/sampler"),
+        toolSuperselect: require("./tool/superselect"),
+        toolSuperselectType: require("./tool/superselect/type"),
+        toolSuperselectVector: require("./tool/superselect/vector"),
+        toolType: require("./tool/type"),
         tools: require("./tools"),
         transform: require("./transform"),
         type: require("./type"),
