+++ src/js/models/tools/ellipse.js
@@ -23,96 +23,52 @@
 
 define(function (require, exports, module) {
     "use strict";
-
-    var Promise = require("bluebird");
-
+  
     var util = require("adapter").util,
-        descriptor = require("adapter").ps.descriptor,
-        toolLib = require("adapter").lib.tool,
         OS = require("adapter").os,
-        UI = require("adapter").ps.ui,
-        vectorMaskLib = require("adapter").lib.vectorMask;
+        UI = require("adapter").ps.ui;
 
     var Tool = require("js/models/tool"),
-        toolActions = require("js/actions/tools"),
         EventPolicy = require("js/models/eventpolicy"),
         KeyboardEventPolicy = EventPolicy.KeyboardEventPolicy,
-        utilShortcuts = require("js/util/shortcuts");
-
-    /**
-     * Sets the tool into either path or shape mode and calls the approprate PS actions based on that mode
-     *
-     * @private
-     */
-    var _selectHandler = function () {
-        var toolStore = this.flux.store("tool"),
-            vectorMode = toolStore.getVectorMode(),
-            toolMode = toolLib.toolModes.SHAPE,
-            firstLaunch = true;
-
-        if (vectorMode) {
-            toolMode = toolLib.toolModes.PATH;
-        }
-
-        var setObj = toolLib.setShapeToolMode(toolMode);
-
-        var setPromise = descriptor.batchPlayObjects([setObj]);
-
-        if (!vectorMode && firstLaunch) {
-            var defaultPromise = this.transfer(toolActions.installShapeDefaults,
-                "rectangleTool");
-
-            firstLaunch = false;
-            return Promise.join(defaultPromise, setPromise);
-        } else if (!vectorMode) {
-            return Promise.join(setPromise);
-        } else {
-            return setPromise
-            .then(function () {
-                return UI.setSuppressTargetPaths(false);
-            })
-            .then(function () {
-                return descriptor.playObject(vectorMaskLib.activateVectorMaskEditing());
-            });
-        }
-    };
+        shortcuts = require("js/util/shortcuts");
 
     /**
      * @implements {Tool}
      * @constructor
      */
-    var RectangleTool = function () {
+    var EllipseTool = function () {
         var shiftUKeyPolicy = new KeyboardEventPolicy(UI.policyAction.PROPAGATE_TO_BROWSER,
-                OS.eventKind.KEY_DOWN, { shift: true }, utilShortcuts.GLOBAL.TOOLS.SHAPE),
+                OS.eventKind.KEY_DOWN, { shift: true }, shortcuts.GLOBAL.TOOLS.SHAPE),
             escapeKeyPolicy = new KeyboardEventPolicy(UI.policyAction.PROPAGATE_TO_BROWSER,
                 OS.eventKind.KEY_DOWN, null, OS.eventKeyCode.ESCAPE),
             deleteKeyPolicy = new KeyboardEventPolicy(UI.policyAction.PROPAGATE_TO_BROWSER,
                 OS.eventKind.KEY_DOWN, null, OS.eventKeyCode.DELETE),
             backspaceKeyPolicy = new KeyboardEventPolicy(UI.policyAction.PROPAGATE_TO_BROWSER,
                 OS.eventKind.KEY_DOWN, null, OS.eventKeyCode.BACKSPACE);
+        
+        Tool.call(this, "ellipse", "Ellipse", "ellipseTool", "toolEllipse.select");
 
-        Tool.call(this, "rectangle", "Rectangle", "rectangleTool", _selectHandler);
-       
         this.keyboardPolicyList = [shiftUKeyPolicy, deleteKeyPolicy, backspaceKeyPolicy, escapeKeyPolicy];
-        this.activationKey = utilShortcuts.GLOBAL.TOOLS.RECTANGLE;
+        this.activationKey = shortcuts.GLOBAL.TOOLS.ELLIPSE;
         this.handleVectorMaskMode = true;
     };
-    util.inherits(RectangleTool, Tool);
+    util.inherits(EllipseTool, Tool);
 
     /**
      * Handler for keydown events, installed when the tool is active.
      *
      * @param {CustomEvent} event
      */
-    RectangleTool.prototype.onKeyDown = function (event) {
+    EllipseTool.prototype.onKeyDown = function (event) {
         var flux = this.getFlux(),
             toolStore = flux.store("tool"),
             detail = event.detail;
 
-        if (detail.keyChar === utilShortcuts.GLOBAL.TOOLS.SHAPE && detail.modifiers.shift) {
-            flux.actions.tools.select(toolStore.getToolByID("ellipse"));
+        if (detail.keyChar === shortcuts.GLOBAL.TOOLS.SHAPE && detail.modifiers.shift) {
+            flux.actions.tools.select(toolStore.getToolByID("rectangle"));
         }
-        
+
         if (toolStore.getVectorMode() && detail.keyCode === OS.eventKeyCode.ESCAPE) {
             flux.actions.tools.changeVectorMaskMode(false);
         }
@@ -122,8 +78,5 @@ define(function (require, exports, module) {
             flux.actions.mask.handleDeleteVectorMask();
         }
     };
-
-    RectangleTool.prototype.toolMode = toolLib.shapeVectorTool;
-
-    module.exports = RectangleTool;
+    module.exports = EllipseTool;
 });
