+++ src/js/models/tools/pen.js
@@ -1,0 +1,91 @@
+/*
+ * Copyright (c) 2014 Adobe Systems Incorporated. All rights reserved.
+ *  
+ * Permission is hereby granted, free of charge, to any person obtaining a
+ * copy of this software and associated documentation files (the "Software"), 
+ * to deal in the Software without restriction, including without limitation 
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense, 
+ * and/or sell copies of the Software, and to permit persons to whom the 
+ * Software is furnished to do so, subject to the following conditions:
+ *  
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *  
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER 
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING 
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER 
+ * DEALINGS IN THE SOFTWARE.
+ * 
+ */
+
+define(function (require, exports, module) {
+    "use strict";
+
+    var util = require("adapter").util,
+        OS = require("adapter").os,
+        UI = require("adapter").ps.ui;
+
+    var Tool = require("js/models/tool"),
+        EventPolicy = require("js/models/eventpolicy"),
+        KeyboardEventPolicy = EventPolicy.KeyboardEventPolicy,
+        shortcutUtil = require("js/util/shortcuts");
+
+    /**
+     * @implements {Tool}
+     * @constructor
+     */
+    var PenTool = function () {
+        Tool.call(this, "pen", "Pen", "penTool");
+
+        var backspaceKeyPolicy = new KeyboardEventPolicy(UI.policyAction.PROPAGATE_TO_BROWSER,
+                OS.eventKind.KEY_DOWN, null, OS.eventKeyCode.BACKSPACE),
+            deleteKeyPolicy = new KeyboardEventPolicy(UI.policyAction.PROPAGATE_TO_BROWSER,
+                OS.eventKind.KEY_DOWN, null, OS.eventKeyCode.DELETE),
+            escapeKeyPolicy = new KeyboardEventPolicy(UI.policyAction.PROPAGATE_TO_BROWSER,
+                OS.eventKind.KEY_DOWN, null, OS.eventKeyCode.ESCAPE),
+            arrowUpKeyPolicy = new KeyboardEventPolicy(UI.policyAction.PROPAGATE_BY_FOCUS,
+                OS.eventKind.KEY_DOWN, null, OS.eventKeyCode.ARROW_UP),
+            arrowRightKeyPolicy = new KeyboardEventPolicy(UI.policyAction.PROPAGATE_BY_FOCUS,
+                OS.eventKind.KEY_DOWN, null, OS.eventKeyCode.ARROW_RIGHT),
+            arrowDownKeyPolicy = new KeyboardEventPolicy(UI.policyAction.PROPAGATE_BY_FOCUS,
+                OS.eventKind.KEY_DOWN, null, OS.eventKeyCode.ARROW_DOWN),
+            arrowLeftKeyPolicy = new KeyboardEventPolicy(UI.policyAction.PROPAGATE_BY_FOCUS,
+                OS.eventKind.KEY_DOWN, null, OS.eventKeyCode.ARROW_LEFT);
+            
+        this.keyboardPolicyList = [
+            backspaceKeyPolicy,
+            deleteKeyPolicy,
+            escapeKeyPolicy,
+            arrowUpKeyPolicy, // Arrow keys for nudging points
+            arrowRightKeyPolicy,
+            arrowDownKeyPolicy,
+            arrowLeftKeyPolicy
+        ];
+        
+        this.selectHandler = "toolPen.select";
+        this.deselectHandler = "toolPen.deselect";
+        this.activationKey = shortcutUtil.GLOBAL.TOOLS.PEN;
+        this.handleVectorMaskMode = true;
+    };
+    util.inherits(PenTool, Tool);
+
+    /**
+     * Handler for keydown events, installed when the tool is active.
+     *
+     * @param {CustomEvent} event
+     */
+    PenTool.prototype.onKeyDown = function (event) {
+        var flux = this.getFlux(),
+            toolStore = flux.store("tool"),
+            detail = event.detail;
+  
+        if (toolStore.getVectorMode() && detail.keyCode === OS.eventKeyCode.ESCAPE) {
+            flux.actions.tools.changeVectorMaskMode(false);
+        }
+    };
+
+    module.exports = PenTool;
+});
