+++ src/js/models/tools/rectangle.js
@@ -23,96 +23,53 @@
 
 define(function (require, exports, module) {
     "use strict";
-  
-    var Promise = require("bluebird");
 
     var util = require("adapter").util,
-        descriptor = require("adapter").ps.descriptor,
         toolLib = require("adapter").lib.tool,
         OS = require("adapter").os,
-        UI = require("adapter").ps.ui,
-        vectorMaskLib = require("adapter").lib.vectorMask;
+        UI = require("adapter").ps.ui;
 
     var Tool = require("js/models/tool"),
-        toolActions = require("js/actions/tools"),
         EventPolicy = require("js/models/eventpolicy"),
         KeyboardEventPolicy = EventPolicy.KeyboardEventPolicy,
-        shortcuts = require("js/util/shortcuts");
-
-    /**
-     * Sets the tool into either path or shape mode and calls the approprate PS actions based on that mode
-     *
-     * @private
-     */
-    var _selectHandler = function () {
-        var toolStore = this.flux.store("tool"),
-            vectorMode = toolStore.getVectorMode(),
-            toolMode = toolLib.toolModes.SHAPE,
-            firstLaunch = true;
-
-        if (vectorMode) {
-            toolMode = toolLib.toolModes.PATH;
-        }
-
-        var setObj = toolLib.setShapeToolMode(toolMode);
-
-        var setPromise = descriptor.playObject(setObj);
-
-        if (!vectorMode && firstLaunch) {
-            var defaultPromise = this.transfer(toolActions.installShapeDefaults,
-                "ellipseTool");
-
-            firstLaunch = false;
-            return Promise.join(defaultPromise, setPromise);
-        } else if (!vectorMode) {
-            return setPromise;
-        } else {
-            return setPromise
-            .then(function () {
-                return UI.setSuppressTargetPaths(false);
-            })
-            .then(function () {
-                return descriptor.playObject(vectorMaskLib.activateVectorMaskEditing());
-            });
-        }
-    };
+        utilShortcuts = require("js/util/shortcuts");
 
     /**
      * @implements {Tool}
      * @constructor
      */
-    var EllipseTool = function () {
+    var RectangleTool = function () {
         var shiftUKeyPolicy = new KeyboardEventPolicy(UI.policyAction.PROPAGATE_TO_BROWSER,
-                OS.eventKind.KEY_DOWN, { shift: true }, shortcuts.GLOBAL.TOOLS.SHAPE),
+                OS.eventKind.KEY_DOWN, { shift: true }, utilShortcuts.GLOBAL.TOOLS.SHAPE),
             escapeKeyPolicy = new KeyboardEventPolicy(UI.policyAction.PROPAGATE_TO_BROWSER,
                 OS.eventKind.KEY_DOWN, null, OS.eventKeyCode.ESCAPE),
             deleteKeyPolicy = new KeyboardEventPolicy(UI.policyAction.PROPAGATE_TO_BROWSER,
                 OS.eventKind.KEY_DOWN, null, OS.eventKeyCode.DELETE),
             backspaceKeyPolicy = new KeyboardEventPolicy(UI.policyAction.PROPAGATE_TO_BROWSER,
                 OS.eventKind.KEY_DOWN, null, OS.eventKeyCode.BACKSPACE);
-        
-        Tool.call(this, "ellipse", "Ellipse", "ellipseTool", _selectHandler);
 
+        Tool.call(this, "rectangle", "Rectangle", "rectangleTool", "toolRectangle.select");
+       
         this.keyboardPolicyList = [shiftUKeyPolicy, deleteKeyPolicy, backspaceKeyPolicy, escapeKeyPolicy];
-        this.activationKey = shortcuts.GLOBAL.TOOLS.ELLIPSE;
+        this.activationKey = utilShortcuts.GLOBAL.TOOLS.RECTANGLE;
         this.handleVectorMaskMode = true;
     };
-    util.inherits(EllipseTool, Tool);
+    util.inherits(RectangleTool, Tool);
 
     /**
      * Handler for keydown events, installed when the tool is active.
      *
      * @param {CustomEvent} event
      */
-    EllipseTool.prototype.onKeyDown = function (event) {
+    RectangleTool.prototype.onKeyDown = function (event) {
         var flux = this.getFlux(),
             toolStore = flux.store("tool"),
             detail = event.detail;
 
-        if (detail.keyChar === shortcuts.GLOBAL.TOOLS.SHAPE && detail.modifiers.shift) {
-            flux.actions.tools.select(toolStore.getToolByID("rectangle"));
+        if (detail.keyChar === utilShortcuts.GLOBAL.TOOLS.SHAPE && detail.modifiers.shift) {
+            flux.actions.tools.select(toolStore.getToolByID("ellipse"));
         }
-
+        
         if (toolStore.getVectorMode() && detail.keyCode === OS.eventKeyCode.ESCAPE) {
             flux.actions.tools.changeVectorMaskMode(false);
         }
@@ -122,5 +79,8 @@ define(function (require, exports, module) {
             flux.actions.mask.handleDeleteVectorMask();
         }
     };
-    module.exports = EllipseTool;
+
+    RectangleTool.prototype.toolMode = toolLib.shapeVectorTool;
+
+    module.exports = RectangleTool;
 });
